Separate benchmarking util from tests
